% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SDA.R
\name{calculate_soil_depth_NRCS}
\alias{calculate_soil_depth_NRCS}
\title{Determine soil depth for \var{NRCS} soil data}
\usage{
calculate_soil_depth_NRCS(
  x,
  target_site_ids,
  restrict_by_ec_or_ph = TRUE,
  var_site_id = "COKEY",
  var_horizon = "Horizon_No",
  var_horizon_lower_depth = "hzdepb_r",
  var_restrictions = c("Horizon_depth", "RootZoneRestriction_depth", "Bedrock_depth"),
  var_soiltexture = c("sand", "clay", "silt")
)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.
Soil horizons/layers are organized in rows
and soil texture variables in columns.}

\item{target_site_ids}{A vector. The unique location identifiers, e.g.,
\var{cokey} values, for which soil depth is to be determined.}

\item{restrict_by_ec_or_ph}{A logical value. Include depth restrictions
by \code{ph <= 3.5} or \code{ec >= 16}. This option requires additional
variables.}

\item{var_site_id}{A character string. The column name of \code{x} which
contains the unique location identifiers.}

\item{var_horizon}{A character string. The column name of \code{x} which
contains the horizon/layer numbers where the shallowest horizon/layer
is number one.}

\item{var_horizon_lower_depth}{A character string. The column name of
\code{x} that contains the lower depth limit of horizons/layers.}

\item{var_restrictions}{A vector of character strings. The column names of
\code{x} that contain depth restrictions, e.g., bedrock.}

\item{var_soiltexture}{A vector of character strings. The column names of
\code{x} that contain the three soil texture variables sand, clay,
and silt.}
}
\value{
A \code{data.frame} with at least three columns. Each row represents
  one value of \code{target_cokeys}. The columns are: \itemize{
    \item \var{N_horizons}: The number of soil horizons/layers.
    \item \var{SoilDepth_cm}: The soil depth in centimeters.
    \item \var{depth_Lx}: The lower depth of soil horizon/layer \var{x}.
      Note: \var{depth_L1} may vary from \code{x[, "hzdepb_r"]}.
  }
}
\description{
Determine soil depth for \var{NRCS} soil data
}
\examples{
\dontrun{
if (curl::has_internet()) {
  x <- fetch_soils_from_NRCS_SDA(mukeys_unique = c(471168, 1606800))

  calculate_soil_depth_NRCS(x, restrict_by_ec_or_ph = FALSE)

  x2 <- cbind(x, organic = is_NRCS_horizon_organic(x))
  calculate_soil_depth_NRCS(x2, restrict_by_ec_or_ph = TRUE)
}
}

}
\references{
Code based on \var{CalcRZDepth()} version 2020-08-31:
  \url{https://github.com/ncss-tech/SoilDataDevelopmentToolbox/blob/master/SDA_Valu2Table.py}
  Note: currently ignores "dense" layer restrictions
}
