% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SOLUS100.R
\name{check_SOLUS100}
\alias{check_SOLUS100}
\title{Check local copy of \code{SOLUS100}}
\usage{
check_SOLUS100(
  path = ".",
  vars = c("resdept_cm", "dbovendry", "fragvol", "sandtotal", "silttotal", "claytotal",
    "soc"),
  depths = depth_profile_SOLUS100(),
  stat = c("p", "l", "h", "rpi")
)
}
\arguments{
\item{path}{A character string. The path to the local copy of \code{SOLUS100}.}

\item{vars}{A vector of variable names. See \code{\link[=variables_SOLUS100]{variables_SOLUS100()}}}

\item{depths}{Soil depths in centimeters from surface.
See \code{\link[=depth_profile_SOLUS100]{depth_profile_SOLUS100()}}}

\item{stat}{A vector of character strings. See Nauman et al. 2024}
}
\description{
Check local copy of \code{SOLUS100}
}
\examples{
dir_tmp <- tempdir()
has_solus <- check_SOLUS100(dir_tmp, vars = "resdept_cm")

\dontrun{
if (curl::has_internet()) {
  fns_solus <- download_SOLUS100(dir_tmp, vars = "resdept_cm")
  files_solus <- file.path(dir_tmp, fns_solus)
  terra::plot(terra::rast(files_solus))
  has_solus <- check_SOLUS100(dir_tmp, vars = "resdept_cm")

  unlink(files_solus) # clean up
}
}

}
\references{
Nauman, T. 2024.
Data from: Soil Landscapes of the United States 100-meter (\code{SOLUS100})
soil property maps project repository. Ag Data Commons.
\url{https://doi.org/10.15482/USDA.ADC/25033856.V1}.
}
\seealso{
\code{\link[=download_SOLUS100]{download_SOLUS100()}}
}
