% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_Miller1998_CONUSSoils.R
\name{conussoil}
\alias{conussoil}
\alias{miller1998}
\title{Functionality to use the \var{CONUSSoil} soil dataset
for \pkg{SOILWAT2} applications}
\arguments{
\item{path}{A character string. The path to the local copy of the
\var{CONUSSoil} folder hierarchy, e.g.,
\code{dirname(prepare_script_for_Miller1998_CONUSSoil())}.}

\item{vars}{A vector of character strings. The requested variables as
used by \var{CONUSSoil}; see Miller et al. 1998.}

\item{lower_limits_by_vars}{A named numeric vector. The names correspond
to \code{vars} and values represent the lower limits that are conditioned/
masked out, i.e., set to \code{NA}.}

\item{digits}{An integer value. The number of digits to which soil texture
variables are rounded. Skip rounding if \code{NA} or \code{NULL}.}

\item{verbose}{A logical value.}
}
\description{
Functionality to use the \var{CONUSSoil} soil dataset
for \pkg{SOILWAT2} applications
}
\references{
Miller, D. A., and R. A. White. 1998. A conterminous United States
 multilayer soil characteristics dataset for regional climate and
 hydrology modeling. Earth Interactions 2:1-26.
 \url{http://www.soilinfo.psu.edu/index.cgi?soil_data&conus}
}
\seealso{
\code{\link{extract_soils_Miller1998_CONUSSoil}}
}
