% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_Miller1998_CONUSSoils.R
\name{create_conditioned_Miller1998_CONUSSoil}
\alias{create_conditioned_Miller1998_CONUSSoil}
\title{Mask out unrealistic variable values in \var{CONUSSoil}}
\usage{
create_conditioned_Miller1998_CONUSSoil(
  path,
  vars = c("rockdepm", "rockvol", "bd", "sand", "clay", "silt"),
  lower_limits_by_vars = c(rockdepm = 0, rockvol = 0, bd = 30, sand = 0, clay = 0, silt =
    0)
)
}
\arguments{
\item{path}{A character string. The path to the local copy of the
\var{CONUSSoil} folder hierarchy, e.g.,
\code{dirname(prepare_script_for_Miller1998_CONUSSoil())}.}

\item{vars}{A vector of character strings. The requested variables as
used by \var{CONUSSoil}; see Miller et al. 1998.}

\item{lower_limits_by_vars}{A named numeric vector. The names correspond
to \code{vars} and values represent the lower limits that are conditioned/
masked out, i.e., set to \code{NA}.}
}
\value{
This function is called for its side effects of creating
  \var{GeoTIFFs} if they don't already exist.
  It returns invisibly a logical vector for each value of \code{vars} where
  \code{TRUE} indicates that the conditioned file already exists or was
  successfully created.
}
\description{
This function creates new \var{CONUSSoil-GeoTIFFs} where
values below a limit are masked out (set as missing). This can be helpful
when aggregating values from multiple grid cells for problematic variables
such as bulk density (\var{bd})
for which some value ranges are unrealistic for representing soils .
}
