% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SOLUS100.R
\name{depth_profile_SOLUS100}
\alias{depth_profile_SOLUS100}
\title{List of soil depths available from \code{SOLUS100}}
\usage{
depth_profile_SOLUS100()
}
\description{
List of soil depths available from \code{SOLUS100}
}
