% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SOLUS100.R
\name{download_SOLUS100}
\alias{download_SOLUS100}
\title{Download soil layers from \code{SOLUS100}}
\usage{
download_SOLUS100(
  path = ".",
  vars = c("resdept_cm", "dbovendry", "fragvol", "sandtotal", "silttotal", "claytotal",
    "soc"),
  depths = depth_profile_SOLUS100(),
  stat = c("p", "l", "h", "rpi"),
  url_solus100 = "https://storage.googleapis.com/solus100pub/",
  verbose = FALSE
)
}
\arguments{
\item{path}{A character string. The path to the local copy of \code{SOLUS100}.}

\item{vars}{A vector of variable names. See \code{\link[=variables_SOLUS100]{variables_SOLUS100()}}}

\item{depths}{Soil depths in centimeters from surface.
See \code{\link[=depth_profile_SOLUS100]{depth_profile_SOLUS100()}}}

\item{stat}{A vector of character strings. See Nauman et al. 2024}

\item{url_solus100}{The \code{URL} to the \code{SOLUS100} data repository.
See Nauman et al. 2024}

\item{verbose}{A logical value.}
}
\value{
File names to local copies of requested soil layers.
}
\description{
Download soil layers from \code{SOLUS100}
}
\examples{
\dontrun{
if (curl::has_internet()) {
  dir_tmp <- tempdir()
  fsolus <- download_SOLUS100(dir_tmp, vars = "resdept_cm")
  terra::plot(terra::rast(file.path(dir_tmp, fsolus)))
  unlink(file.path(dir_tmp, fsolus)) # clean up
}
}

}
\references{
Nauman, T. 2024.
Data from: Soil Landscapes of the United States 100-meter (\code{SOLUS100})
soil property maps project repository. Ag Data Commons.
\url{https://doi.org/10.15482/USDA.ADC/25033856.V1}.
}
\seealso{
\code{\link[=check_SOLUS100]{check_SOLUS100()}}
}
