% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_KSSL.R
\name{extract_soils_KSSL}
\alias{extract_soils_KSSL}
\title{Extract soil information from the \code{KSSL} soil dataset via \code{SDA}
for \pkg{SOILWAT2} applications}
\usage{
extract_soils_KSSL(
  x,
  what = c("pedlabsampnum", "upedonid", "pedon_key"),
  replace_missing_fragvol_with_zero = FALSE,
  replace_missing_om_with_zero = FALSE,
  estimate_missing_bulkdensity = FALSE,
  fix_depths = FALSE,
  digits = 3L,
  verbose = FALSE
)
}
\arguments{
\item{x}{A vector of character strings. Soil identifier(s);
the type of identifier is specified by \code{what}.}

\item{what}{A character string.
\code{"pedlabsampnum"} (Laboratory Pedon ID), \code{"upedonid"} (User Pedon ID),
or \code{"pedon_key"} (Pedon ID).
see \code{\link[=variables_KSSL]{variables_KSSL()}} for implemented variables.}

\item{replace_missing_fragvol_with_zero}{A logical value. Replace missing
values of coarse fragments with 0.}

\item{replace_missing_om_with_zero}{A character value. Replace missing
values of organic matter with 0.}

\item{estimate_missing_bulkdensity}{A logical value. Estimate missing
bulk density values from saturated water content and gravel volume.
See \code{\link[rSW2data]{estimate_bulkdensity}}.}

\item{fix_depths}{A logical value. Fix depth issues, see \code{\link[=fixHorizonDepths]{fixHorizonDepths()}}}

\item{digits}{An integer value. The number of digits to which soil texture
variables are rounded. Skip rounding if \code{NA} or \code{NULL}.}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the \code{KSSL} soil dataset via \code{SDA}
for \pkg{SOILWAT2} applications
}
\examples{
\dontrun{
if (curl::has_internet()) {
x <- c("04N0873", "10N1336", "10N1239", "02N0184")
res <- extract_soils_KSSL(x = x, what = "pedlabsampnum", fix_depths = TRUE)
}
}

}
\seealso{
\code{\link[soilDB:fetchLDM]{soilDB::fetchLDM()}}
}
