% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_Miller1998_CONUSSoils.R
\name{extract_soils_Miller1998_CONUSSoil}
\alias{extract_soils_Miller1998_CONUSSoil}
\title{Extract soil information from the \var{CONUSSoil} soil dataset
for \pkg{SOILWAT2} applications}
\usage{
extract_soils_Miller1998_CONUSSoil(
  x,
  crs = 4326,
  path = ".",
  vars = c("bd", "rockvol", "sand", "clay", "silt"),
  lower_limits_by_vars = c(bd = 30, rockvol = 0, sand = 0, clay = 0, silt = 0),
  replace_missing_fragvol_with_zero = c("none", "all", "at_surface"),
  impute = FALSE,
  digits = 3L,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \code{\link[terra:SpatVector-class]{terra::SpatVector}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-class]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{path}{A character string. The path to the local copy of the
\var{CONUSSoil} folder hierarchy, e.g.,
\code{dirname(prepare_script_for_Miller1998_CONUSSoil())}.}

\item{vars}{A vector of character strings. The requested variables as
used by \var{CONUSSoil}; see Miller et al. 1998.}

\item{lower_limits_by_vars}{A named numeric vector. The names correspond
to \code{vars} and values represent the lower limits that are conditioned/
masked out, i.e., set to \code{NA}.}

\item{replace_missing_fragvol_with_zero}{A character string. Method
indicating how missing/null values of rock/gravel fragment fractions
should be interpreted;
passed to \code{\link[rSW2data]{set_missing_soils_to_value}}.
The options are one of
\describe{
  \item{\var{"all"}}{
    Missing/null values of rack/gravel fragments are
    replaced by 0 %.
    See also argument \var{nullFragsAreZero} of
    function \code{\link[soilDB]{fetchSDA}}.
  }
  \item{\var{"at_surface"}}{
    Missing/null values of rack/gravel fragments are
    replaced by 0 % in the shallowest horizon only.
    Note, remaining missing values in deeper horizons
    can subsequently be imputed by argument \code{impute}.
  }
  \item{\var{"none"}}{
     Missing/null values remain unmodified unless
     argument \code{impute} is activated.
  }
}}

\item{impute}{A logical value. Impute missing values with a
shallow-depth value carried deeper approach (in analogy to \var{LOCF}).
Consequently, missing values in the shallowest horizon are not imputed.
See \code{\link[rSW2utils]{impute_df}}.}

\item{digits}{An integer value. The number of digits to which soil texture
variables are rounded. Skip rounding if \code{NA} or \code{NULL}.}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the \var{CONUSSoil} soil dataset
for \pkg{SOILWAT2} applications
}
\section{Notes}{
 A local copy of \var{CONUSSoil} is required. The function
  \code{\link{prepare_script_for_Miller1998_CONUSSoil}} creates a script
  that can be used to download and prepare \var{CONUSSoil} files.
}

\examples{
script_to_download_conussoil <- prepare_script_for_Miller1998_CONUSSoil()

## Execute script to download data
## (or set `path_conussoil` to your local copy)
path_conussoil <- dirname(script_to_download_conussoil)

## Mask out unrealistic variable values
has_CONUSSoil_cond <- create_conditioned_Miller1998_CONUSSoil(
  path = path_conussoil
)

## Check that we have CONUSSoil data
has_CONUSSoil <- isTRUE(all(
  check_Miller1998_CONUSSoil(path = path_conussoil)
))

if (has_CONUSSoil) {

  locations <- matrix(
    data = c(-120.1286878, -111.8511136, 39.8182913, 36.9047396),
    nrow = 2
  )

  res <- extract_soils_Miller1998_CONUSSoil(
    x = locations,
    path = path_conussoil,
    verbose = TRUE
  )
}

# Clean up example
unlink(script_to_download_conussoil)


}
\references{
Miller, D. A., and R. A. White. 1998. A conterminous United States
 multilayer soil characteristics dataset for regional climate and
 hydrology modeling. Earth Interactions 2:1-26.
 \url{http://www.soilinfo.psu.edu/index.cgi?soil_data&conus}
}
\seealso{
\code{\link[raster]{extract}}
}
