% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SOLUS100.R
\name{extract_soils_SOLUS100}
\alias{extract_soils_SOLUS100}
\title{Extract soil information from the \code{SOLUS100} soil dataset
for \pkg{SOILWAT2} applications}
\usage{
extract_soils_SOLUS100(
  x,
  crs = 4326,
  vars = c("dbovendry", "fragvol", "sandtotal", "silttotal", "claytotal", "soc"),
  var_depth = "resdept_cm",
  depths = depth_profile_SOLUS100(),
  stat = c("p", "l", "h", "rpi"),
  path = ".",
  method_vertical = c("asis", "interpolate_by_layer"),
  requested_layer_depths = NULL,
  method_horizontal = c("asis", "fix_with_buffer"),
  fix_criteria = list(dbovendry = list(op = "<", value = 0.6), texture = list(op = "<",
    value = 0.5)),
  buffer_m = NULL,
  fun = NULL,
  na.rm = TRUE,
  digits = 3L,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \code{\link[terra:SpatVector-class]{terra::SpatVector}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-class]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{vars}{A vector of variable names. See \code{\link[=variables_SOLUS100]{variables_SOLUS100()}}}

\item{var_depth}{A character string or \code{NULL}.
If \code{NULL}, then soil depth is determined based on
(i) available data layers \code{depths} (up to 201 cm; see Nauman et al. 2024),
(ii) depth of non-missing values in extracted \code{vars} (which may be 0).
Otherwise, soil depth is the extracted value of the \code{SOLUS100}
variable \code{var_depth}.
Soil depth is set to 0 if is missing or all \code{vars} values are missing.}

\item{depths}{Soil depths in centimeters from surface.
See \code{\link[=depth_profile_SOLUS100]{depth_profile_SOLUS100()}}}

\item{stat}{A vector of character strings. See Nauman et al. 2024}

\item{path}{A character string. The path to the local copy of \code{SOLUS100}.}

\item{method_vertical}{A character string that is
\code{"asis"} or \code{"interpolate_by_layer"}:
(i) method \code{"asis"} extracts the values as provided by \code{SOLUS100},
i.e., as point estimates at specified depths
(ii) method \code{"interpolate_by_layer"} interpolates extracted values
for each centimeter depth increment and averages
the interpolated values across requested soil layers
(see \code{requested_layer_depths}).
Soils properties for a soil with a missing depth value or a
depth of less than 1 centimeter are set to \code{NA}.}

\item{requested_layer_depths}{An integer vector
(used if \code{method_vertical = "interpolate_by_layer"}).
Soil depths (in centimeters) at lower layer boundaries used for output
If \code{NULL}, then layer boundaries are assumed to be \code{c(depths, 201)}.}

\item{method_horizontal}{A character string.
Method that determines the extraction approach across grid cells:
(i) values are extracted using arguments
\code{buffer_m}, \code{fun}, and \code{na.rm}
and are returned \code{"asis"} or
(ii) values are extracted for point locations,
i.e., temporarily setting \code{buffer_m = NULL}; then,
sites with problematic values (as determined by \code{fix_criteria})
are extracted again under \code{"fix_with_buffer"} based on
\code{buffer_m}, \code{fun}, and \code{na.rm}}

\item{fix_criteria}{A named list
(used if \code{method_horizontal = "fix_with_buffer"}).
Names match values of \code{vars} or one of the names can be \code{"texture"}
(in which case the criterion is applied to sand, clay, and silt).
Each element is applied to the variable that corresponds to the name
and is used to determine whether a site has problematic values.
Elements themselves are a named list with two sub-elements
\code{"op"} for the relationship operator, e.g., \code{"<"}, and
\code{"value"} for the value to compare against. See examples.}

\item{buffer_m}{A numeric value. The radius of a buffer around each point
from which to extract cell values and across which \code{fun} is applied.
Set to \code{NULL} to extract \code{SOLUS100} gridcell values at point locations.}

\item{fun}{A function or a named list containing functions
(used if \code{method_horizontal = "fix_with_buffer"}).
Names match values of \code{vars} or one of the names can be \code{"texture"}
(in which case the criterion is applied to sand, clay, and silt).
The function(s) summarize(s) extracted values if more than one value
is extracted per location (based on \code{buffer_m}).}

\item{na.rm}{A logical value. Passed to \code{fun}.}

\item{digits}{An integer value. The number of digits to which soil texture
variables are rounded. Skip rounding if \code{NA} or \code{NULL}.}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the \code{SOLUS100} soil dataset
for \pkg{SOILWAT2} applications
}
\section{Notes}{
 A local copy of \code{SOLUS100} is required. The function
\code{\link[=download_SOLUS100]{download_SOLUS100()}} can be used to download \code{SOLUS100} files.
}

\examples{
\dontrun{
if (curl::has_internet()) {
path_solus100 <- tempdir()
req_vars <- c("resdept_cm", "sandtotal")
req_depths <- 0

## Download data
fns_solus100 <- rSW2exter::download_SOLUS100(
  path = path_solus100,
  vars = req_vars,
  depths = req_depths
)

## Check that we have SOLUS100 data
has_SOLUS100 <- isTRUE(all(
  check_SOLUS100(
    path = path_solus100,
    vars = req_vars,
    depths = req_depths
  )
))

if (has_SOLUS100) {
  locations <- matrix(
    data = c(-120.1286878, -111.8511136, 39.8182913, 36.9047396),
    nrow = 2
  )

  ## Extract gridcell values at point locations
  res <- extract_soils_SOLUS100(
    x = locations,
    vars = req_vars,
    depths = req_depths,
    path = path_solus100
  )
}

# Clean up example
unlink(file.path(path_solus100, fns_solus100))
}
}

}
\references{
Nauman, T. 2024.
Data from: Soil Landscapes of the United States 100-meter (\code{SOLUS100})
soil property maps project repository. Ag Data Commons.
\url{https://doi.org/10.15482/USDA.ADC/25033856.V1}.
}
\seealso{
\code{\link[terra:extract]{terra::extract()}}
}
