% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_Miller1998_CONUSSoils.R
\name{fetch_soils_from_Miller1998_CONUSSoil}
\alias{fetch_soils_from_Miller1998_CONUSSoil}
\title{Extract soil information from the \var{CONUSSoil} soil dataset}
\usage{
fetch_soils_from_Miller1998_CONUSSoil(
  x,
  crs,
  vars,
  lower_limits_by_vars,
  path,
  verbose
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-classes]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{vars}{A vector of character strings. The requested variables as
used by \var{CONUSSoil}; see Miller et al. 1998.}

\item{lower_limits_by_vars}{A named numeric vector. The names correspond
to \code{vars} and values represent the lower limits that are conditioned/
masked out, i.e., set to \code{NA}.}

\item{path}{A character string. The path to the local copy of the
\var{CONUSSoil} folder hierarchy, e.g.,
\code{dirname(prepare_script_for_Miller1998_CONUSSoil())}.}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the \var{CONUSSoil} soil dataset
}
\section{Notes}{
 This is a function with minimal functionality;
use \code{\link{extract_soils_Miller1998_CONUSSoil}}
for a user-friendly interface.
}

\references{
Miller, D. A., and R. A. White. 1998. A conterminous United States
 multilayer soil characteristics dataset for regional climate and
 hydrology modeling. Earth Interactions 2:1-26.
 \url{http://www.soilinfo.psu.edu/index.cgi?soil_data&conus}
}
