% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SDA.R
\name{fetch_soils_from_NRCS_SDA}
\alias{fetch_soils_from_NRCS_SDA}
\title{Download soil data from \code{NRCS} \code{SDA} web service}
\usage{
fetch_soils_from_NRCS_SDA(
  mukeys_unique,
  bind_params = mukeys_unique,
  sql_template = NA,
  injection_format = "mukey IN (\%s)",
  majcompflag = c("subset", "ignore"),
  only_soilcomp = TRUE,
  chunk_size = 1000L,
  progress_bar = FALSE
)
}
\arguments{
\item{mukeys_unique}{An integer vector with unique \code{mukey} values.
Deprecated; use \code{bind_params} instead.}

\item{bind_params}{A vector or 2-dimensional container with parameter values
to bind to the \code{T-SQL} query via \code{injection_format}.}

\item{sql_template}{A character vector.
A valid \code{T-SQL} query with a \code{"WHERE"} clause so that the code can
inject chunks of \code{bind_params} values via format \code{injection_format}.
If \code{NA}, then the default query is loaded, see examples.}

\item{injection_format}{A character vector that identifies the location
(format specifier) in \code{sql_template} for parameter value injection/binding.}

\item{majcompflag}{A character string. \code{"subset"} keeps
the \code{"WHERE"} clause \code{component.majcompflag = 'Yes'} that is contained in
\code{sql_template}; \code{"ignore"} removes it from the query. Note that
the field \code{"majcompflag"} exists only in the \code{SSURGO} version
of the \code{"component"} table, but not in the \code{STATSGO} version.}

\item{only_soilcomp}{A logical value. If \code{TRUE}, then query restricts
to soil components. If \code{FALSE}, then query includes
all components including \code{"Miscellaneous areas"} and \code{"NOTCOM"}
(not complete) components.}

\item{chunk_size}{An integer value. The size of chunks into which
\code{bind_params} is broken up and looped over for processing.}

\item{progress_bar}{A logical value. Display a progress bar as the code
loops over the chunks?}
}
\value{
A \code{data.frame} according to the specifications of \code{sql} or
\code{NULL} if the query returns empty.
}
\description{
Download soil data from \code{NRCS} \code{SDA} web service
}
\section{Notes}{
 A live internet connection is required to access \code{SDA}.

 This is a function with minimal functionality;
use \code{\link[=extract_soils_NRCS_SDA]{extract_soils_NRCS_SDA()}} for a user-friendly interface.
}

\examples{
\dontrun{
if (curl::has_internet()) {
  # Query soils of dominant component of soil map unit
  fetch_soils_from_NRCS_SDA(bind_params = 67616)

  # As of 2022-March-15, mukey 2479921 contained one "NOTCOM" component
  fetch_soils_from_NRCS_SDA(bind_params = 2479921)
  fetch_soils_from_NRCS_SDA(bind_params = 2479921, only_soilcomp = FALSE)

  sql <- readLines(
    system.file("NRCS", "nrcs_sql_template.sql", package = "rSW2exter")
  )

  fetch_soils_from_NRCS_SDA(bind_params = 67616, sql_template = sql)

  # This will return NULL because -1 is not an existing mukey value
  fetch_soils_from_NRCS_SDA(bind_params = -1, sql_template = sql)

  # Query soils of a specific component of a soil map unit
  sql2 <- readLines(
    system.file("NRCS", "nrcs_sql_template2.sql", package = "rSW2exter")
  )

  fetch_soils_from_NRCS_SDA(
    bind_params = data.frame(mukey = 398856, compname = "Waupaca"),
    sql_template = sql2,
    injection_format = "(VALUES \%s) AS t (mm, cn)"
  )
}
}

}
\seealso{
\code{\link[soilDB:SDA_query]{soilDB::SDA_query()}}
}
