% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SOLUS100.R
\name{fetch_soils_from_SOLUS100}
\alias{fetch_soils_from_SOLUS100}
\title{Extract soil information from the \code{SOLUS100} soil dataset}
\usage{
fetch_soils_from_SOLUS100(
  x,
  vars,
  depths,
  stat,
  path = ".",
  fun = NULL,
  na.rm = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \code{\link[terra:SpatVector-class]{terra::SpatVector}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{vars}{A vector of variable names. See \code{\link[=variables_SOLUS100]{variables_SOLUS100()}}}

\item{depths}{Soil depths in centimeters from surface.
See \code{\link[=depth_profile_SOLUS100]{depth_profile_SOLUS100()}}}

\item{stat}{A vector of character strings. See Nauman et al. 2024}

\item{path}{A character string. The path to the local copy of \code{SOLUS100}.}

\item{fun}{A function. Summarizing gridcell values if more than one value
is extracted per location. See \code{\link[terra:extract]{terra::extract()}}.}

\item{na.rm}{A logical value. Passed to \code{fun}.}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the \code{SOLUS100} soil dataset
}
\section{Notes}{
 This is a function with minimal functionality;
use \code{\link[=extract_soils_SOLUS100]{extract_soils_SOLUS100()}} for a user-friendly interface.
}

\references{
Nauman, T. 2024.
Data from: Soil Landscapes of the United States 100-meter (\code{SOLUS100})
soil property maps project repository. Ag Data Commons.
\url{https://doi.org/10.15482/USDA.ADC/25033856.V1}.
}
