% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SOLUS100.R
\name{filenames_SOLUS100}
\alias{filenames_SOLUS100}
\title{Compose a file name of \code{SOLUS100}}
\usage{
filenames_SOLUS100(vars, depths, stat)
}
\arguments{
\item{vars}{A vector of variable names. See \code{\link[=variables_SOLUS100]{variables_SOLUS100()}}}

\item{depths}{Soil depths in centimeters from surface.
See \code{\link[=depth_profile_SOLUS100]{depth_profile_SOLUS100()}}}

\item{stat}{A vector of character strings. See Nauman et al. 2024}
}
\value{
A vector of file names.
}
\description{
Compose a file name of \code{SOLUS100}
}
