% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_KSSL.R
\name{fixHorizonDepths}
\alias{fixHorizonDepths}
\title{Fix several inconsistencies in horizon depths}
\usage{
fixHorizonDepths(x, vars_depth = c("hzn_top", "hzn_bot"))
}
\arguments{
\item{x}{A data frame with horizon depths and possibly other columns.}

\item{vars_depth}{A vector with two character strings. The column names of
\code{x} with the top horizon depths (first value) and the bottom horizon depths
(second value).}
}
\value{
A copy of \code{x} with possibly updated columns \code{vars_depth} and a new
column \code{"MethodFixedDepth"} that describes the applied fixes.
}
\description{
Fix several inconsistencies in horizon depths
}
\section{Details}{

\itemize{
\item \code{"Hz1Set0cmTop"}: Add 1 cm to bottom-most horizons with a
missing (\code{NA}) bottom depth (see also \code{aqp::accumulateDepths}).
\item \code{"HzkDeepestCreated1cmBottom"}: Set a missing (\code{NA}) top depth of
top-most horizon to 0.
\item \code{"HzkImputedMissingTopDepth"}, \code{"HzkImputedMissingBottomDepth"}:
Set remaining missing (\code{NA}) depth values to non-missing adjacent value.
\item \code{"HzkResetBottomDepthToDeeperTopDepth(old=value)"}: Reset bottom depth
of a horizon \code{k} where the bottom depth is deeper than
the next horizon's (\code{k + 1}) bottom depth to that horizon's top depth.
\item \code{"HzkResetToAverageBottomTopDepths(old=value|value)"}:
Recalculate bottom depth of horizon \code{k} and top depth of horizon \code{k + 1}
if these values disagree but are within the top depth of horizon \code{k} and
the bottom depth of horizon \code{k + 1}.
}
}

\examples{
x <- data.frame(hzn_top = c(NA, 3, 5, 10), hzn_bot = c(3, 5, 10, NA))
fixHorizonDepths(x)

x <- data.frame(hzn_top = c(0, NA, 5, 10), hzn_bot = c(3, 5, 10, 15))
fixHorizonDepths(x)

x <- data.frame(hzn_top = c(0, 0, 0, 0), hzn_bot = c(NA, NA, NA, NA))
fixHorizonDepths(x)

x <- data.frame(hzn_top = c(0, 3, 5, 10), hzn_bot = c(3, 5, 20, 15))
fixHorizonDepths(x)

x <- data.frame(hzn_top = c(0, 3, 7, 10), hzn_bot = c(3, 5, 10, 15))
fixHorizonDepths(x)

}
