% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_Miller1998_CONUSSoils.R
\name{prepare_script_for_Miller1998_CONUSSoil}
\alias{prepare_script_for_Miller1998_CONUSSoil}
\title{Create \var{wget} script to prepare soil data files from \var{CONUSSoil}}
\usage{
prepare_script_for_Miller1998_CONUSSoil(
  path = ".",
  vars = c("rockdepm", "rockvol", "bd", "sand", "clay", "silt")
)
}
\arguments{
\item{path}{A character string. The path to where the local copy of the
\var{CONUSSoil} folder hierarchy and files should be downloaded.}

\item{vars}{A vector of character strings. The requested variables as
used by \var{CONUSSoil}; see Miller et al. 1998.}
}
\value{
(Invisibly) the file path of the generated bash script.
}
\description{
This function only writes out a text file to disk. The user
is responsible to run the script after making sure that \var{wget} is
available and the script is executable.
The script will download the \var{CONUSSoil} files
for the requested variables and convert the \var{ESRI} formatted files to
\var{GeoTIFF}.
}
\section{Notes}{
 this function is not yet implemented!
}

\examples{
fname_wget_miller1998 <- prepare_script_for_Miller1998_CONUSSoil()

## in a shell
## give execute permission if needed: chmod +x <fname_wget_miller1998>
## download data: ./<fname_wget_miller1998>

unlink(fname_wget_miller1998)

}
\references{
Miller, D. A., and R. A. White. 1998. A conterminous United States
 multilayer soil characteristics dataset for regional climate and
 hydrology modeling. Earth Interactions 2:1-26.
 \url{http://www.soilinfo.psu.edu/index.cgi?soil_data&conus}
}
