% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2extern-package.R
\docType{package}
\name{rSW2exter}
\alias{rSW2exter}
\alias{rSW2exter-package}
\title{Package \pkg{rSW2exter}: Access external data as input
 for SOILWAT2 and STEPWAT2 simulations}
\description{
Etymology: `exter` is a Latin adjective meaning "external" or "outer".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrylandEcology/rSW2exter}
  \item Report bugs at \url{https://github.com/DrylandEcology/rSW2exter/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schlaepfer \email{danielrschlaepfer@gmail.com} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

}
