% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SOLUS100.R
\name{variables_SOLUS100}
\alias{variables_SOLUS100}
\title{List of variables available from \code{SOLUS100}}
\usage{
variables_SOLUS100()
}
\description{
List of variables available from \code{SOLUS100}
}
